store: refactored keys to localItems, fixed the bug with repubtime
When creating items, I had assigned the republish time to the expire time as they should be the same according to the paper. BUT due to tests i ofc wanted to try both. Changing republish time was ineffective since the new items repubtime was always created with  respect to the expiretime.

Refactored keys to localItems so that we can republish our saved values to new nodes.

100% coverage and linted.